pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "maturin>=1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ultimate-rotoscopy"
version = "1.0.0"
description = "Ultimate AI-Powered Rotoscopy Application - SAM3, Depth Anything V3, Matte Anything"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ultimate Rotoscopy Team", email = "contact@rotoscopy.ai"}
]
keywords = [
    "rotoscopy", "vfx", "compositing", "matte", "segmentation",
    "depth", "sam3", "depth-anything", "matte-anything", "flame"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "scipy>=1.11.0",
    "scikit-image>=0.21.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.19.0",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "kornia>=0.7.0",
    "openexr>=3.2.0",
    "imath>=3.1.0",
    "imageio>=2.31.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "onnxruntime-gpu>=1.16.0",
    "gradio>=4.0.0",
    "open3d>=0.18.0",
    "trimesh>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
cuda = [
    "cupy-cuda12x>=12.0.0",
]

[project.scripts]
rotoscopy = "ultimate_rotoscopy.cli:main"
rotoscopy-gui = "ultimate_rotoscopy.gui:launch"

[project.urls]
Homepage = "https://github.com/ultimate-rotoscopy/ultimate-rotoscopy"
Documentation = "https://ultimate-rotoscopy.readthedocs.io"
Repository = "https://github.com/ultimate-rotoscopy/ultimate-rotoscopy"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
